requirements.txt
requests
python-dotenv
snowflake-connector-python
streamlit
plotly
pandas
